new app `friends` and some fixes
