+++ backend/friends/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from friends.models import FriendRequest
+
+
+admin.site.register(FriendRequest)
