+++ backend/users/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 
-from .models import User, FriendRequest
+from users.models import User
 
 
 admin.site.register(User)
-admin.site.register(FriendRequest)
