+++ backend/users/urls.py
@@ -1,6 +1,6 @@
-from django.urls import path
+from django.urls import include, path
 
-from . import views
+from users import views
 
 
 urlpatterns = [
@@ -8,8 +8,5 @@ urlpatterns = [
   path('<int:user_id>/', views.get_user, name='get-user'),
   path('me/', views.me, name='me'),
   path('dfa/', views.DFA.as_view(), name='dfa'),
-  path('friends/<int:user_id>/', views.get_user_friends_list, name='friends-list'),
-  path('friends/add/', views.add_friend, name='add-friend'),
-  path('friends/reject/', views.reject_friend, name='reject-friend'),
-  path('friends/remove/', views.remove_friend, name='remove-friend'),
+  path('friends/', include('friends.urls')),
 ]
