+++ backend/users/views.py
@@ -1,7 +1,6 @@
 from django.forms import model_to_dict
 from django.http import HttpRequest, JsonResponse
-from django.shortcuts import get_object_or_404
-from django.views.decorators.http import require_POST, require_GET
+from django.views.decorators.http import require_GET
 from django.utils.decorators import method_decorator
 from django.views.decorators.http import require_GET
 from django.views import View
@@ -12,7 +11,7 @@ import jwt
 import pyotp
 
 from backend import settings
-from users.models import FriendRequest, User
+from users.models import User
 from utils.decorators import need_user
 from random_username.generate import generate_username
 
@@ -108,85 +107,3 @@ class DFA(View):
 def get_user(request: HttpRequest, user_id: int) -> JsonResponse:
   user = User.objects.get(pk=user_id)
   return JsonResponse(model_to_dict(user, exclude=['friends', 'blocked']))
-
-@require_GET
-def get_user_friends_list(request: HttpRequest, user_id: int) -> JsonResponse:
-  user = User.objects.get(pk=user_id)
-  return JsonResponse(list(user.friends.values()), safe=False)
-
-@require_POST
-@jwt_verify
-def add_friend(request: HttpRequest) -> JsonResponse:
-  from_user, to_user = request.payload.get('from_user'), request.payload.get('to_user')
-
-  if None in [from_user, to_user]:
-    return JsonResponse({'error': 'Bad Request', 'message': 'Missing required fields.'}, status=400)
-
-  sender = get_object_or_404(User, pk=from_user)
-  receiver = get_object_or_404(User, pk=to_user)
-
-  try: # has receiver already sent friend request
-    friend_request = FriendRequest.objects.get(sender=receiver, receiver=sender)
-  except FriendRequest.DoesNotExist: # else create it
-    friend_request, created = FriendRequest.objects.get_or_create(sender=sender, receiver=receiver)
-
-    if friend_request.status == 'rejected': # if receiver has rejected it
-      friend_request.status = 'pending'
-      friend_request.save()
-    elif friend_request.status == 'accepted': # if receiver has accepted it
-      if not receiver.friends.contains(sender): # did someone break this friendship
-        friend_request.status = 'pending'
-
-    return JsonResponse(model_to_dict(friend_request))
-
-  if friend_request.status == 'pending': # if you never accepted it
-    friend_request.status = 'accepted'
-    sender.friends.add(receiver)
-  elif friend_request.status == 'accepted': # if you have already accepted it
-    if not receiver.friends.contains(sender): # did someone break this friendship
-      friend_request, created = FriendRequest.objects.get_or_create(sender=sender, receiver=receiver)
-      friend_request.status = 'pending'
-  elif friend_request.status == 'rejected': # if you have already rejected it
-    friend_request, created = FriendRequest.objects.get_or_create(sender=sender, receiver=receiver)
-    friend_request.status = 'pending'
-
-  friend_request.save()
-  print(friend_request)
-
-  return JsonResponse(model_to_dict(friend_request))
-
-@require_POST
-@jwt_verify
-def reject_friend(request: HttpRequest) -> JsonResponse:
-  from_user, to_user = request.payload.get('from_user'), request.payload.get('to_user')
-
-  if None in [from_user, to_user]:
-    return JsonResponse({'error': 'Bad Request', 'message': 'Missing required fields.'}, status=400)
-
-  sender = get_object_or_404(User, pk=from_user)
-  receiver = get_object_or_404(User, pk=to_user)
-
-  friend_request = get_object_or_404(FriendRequest, sender=receiver, receiver=sender)
-
-  if friend_request.status != 'rejected':
-    friend_request.status = 'rejected'
-    friend_request.save()
-
-  print(friend_request)
-
-  return JsonResponse(model_to_dict(friend_request))
-
-@require_POST
-@jwt_verify
-def remove_friend(request: HttpRequest) -> JsonResponse:
-  from_user, to_user = request.payload.get('from_user'), request.payload.get('to_user')
-
-  if None in [from_user, to_user]:
-    return JsonResponse({'error': 'Bad Request', 'message': 'Missing required fields.'}, status=400)
-
-  sender = get_object_or_404(User, pk=from_user)
-  receiver = get_object_or_404(User, pk=to_user)
-
-  sender.friends.remove(receiver)
-
-  return JsonResponse(list(sender.friends.values()), safe=False)
